requirements.txt
# Azure Functions dependencies
azure-functions>=1.17.0

# Additional dependencies for the function app
requests>=2.28.0
